coverartworker: try-out commit about worker&safelywritablefile
