+++ src/sources/metadatasourcetaglib.cpp
@@ -10,10 +10,7 @@
 
 #include "track/taglib/trackmetadata.h"
 #include "util/logger.h"
-
-#if defined(__WINDOWS__)
-#include <Windows.h>
-#endif
+#include "util/safelywritablefile.h"
 
 namespace mixxx {
 
@@ -23,19 +20,6 @@ Logger kLogger("MetadataSourceTagLib");
 
 // TODO(uklotzde): Add a configurable option in the user settings
 const bool kExportTrackMetadataIntoTemporaryFile = true;
-
-// Appended to the original file name of the temporary file used for writing
-const QString kSafelyWritableTempFileSuffix = QStringLiteral("_temp");
-
-// Appended to the original file name for renaming and before deleting this
-// file. Should not be longer than kSafelyWritableTempFileSuffix to avoid
-// potential failures caused by exceeded path length.
-const QString kSafelyWritableOrigFileSuffix = QStringLiteral("_orig");
-
-#if defined(__WINDOWS__)
-const int kWindowsSharingViolationMaxRetries = 5;
-const int kWindowsSharingViolationSleepBeforeNextRetryMillis = 100;
-#endif
 
 // Workaround for missing functionality in TagLib 1.11.x that
 // doesn't support to read text chunks from AIFF files.
@@ -615,276 +599,6 @@ class AiffTagSaver : public TagSaver {
 
     TagLib::RIFF::AIFF::File m_file;
     bool m_modifiedTags;
-};
-
-/**
- * When writing the tags in-place directly into the original file
- * an intermediate failure might corrupt this precious file. For
- * example this might occur if the application crashes or is quit
- * unexpectedly, if the original file becomes unavailable while
- * writing by disconnecting a drive, if the file system is running
- * out of free space, or if an unexpected driver or hardware failure
- * occurs.
- *
- * To reduce the risk of corrupting the original file all write
- * operations are performed on a temporary file that is created
- * as an exact copy of the original file. Only after all write
- * operations have finished successfully the original file is
- * replaced with the temporary file.
- */
-class SafelyWritableFile final {
-  public:
-    SafelyWritableFile(QString origFileName, bool useTemporaryFile) {
-        // Both file names remain uninitialized until all prerequisite operations
-        // in the constructor have been completed successfully. Otherwise failure
-        // to create the temporary file will not be handled correctly!
-        // See also: https://bugs.launchpad.net/mixxx/+bug/1815305
-        DEBUG_ASSERT(m_origFileName.isNull());
-        DEBUG_ASSERT(m_tempFileName.isNull());
-        if (!QFileInfo(origFileName).isWritable()) {
-            kLogger.warning()
-                    << "Failed to prepare file for writing:"
-                    << origFileName
-                    << "is not writable.";
-            // Abort constructor
-            return;
-        }
-        if (useTemporaryFile) {
-            QString tempFileName = origFileName + kSafelyWritableTempFileSuffix;
-            QFile origFile(origFileName);
-            if (!origFile.copy(tempFileName)) {
-                kLogger.warning()
-                        << origFile.errorString()
-                        << "- Failed to clone original into temporary file before writing:"
-                        << origFileName
-                        << "->"
-                        << tempFileName;
-                // Abort constructor
-                return;
-            }
-            QFile tempFile(tempFileName);
-            DEBUG_ASSERT(tempFile.exists());
-            // Both file sizes are expected to be equal after successfully
-            // copying the file contents.
-            VERIFY_OR_DEBUG_ASSERT(origFile.size() == tempFile.size()) {
-                kLogger.warning()
-                        << "Failed to verify size after cloning original into temporary file before writing:"
-                        << origFile.size()
-                        << "<>"
-                        << tempFile.size();
-                // Cleanup
-                if (tempFile.exists() && !tempFile.remove()) {
-                    kLogger.warning()
-                            << tempFile.errorString()
-                            << "- Failed to remove temporary file:"
-                            << tempFileName;
-                }
-                // Abort constructor
-                return;
-            }
-            // Successfully cloned original into temporary file for writing - finish initialization
-            m_origFileName = std::move(origFileName);
-            m_tempFileName = std::move(tempFileName);
-        } else {
-            // Directly write into original file - finish initialization
-            m_origFileName = std::move(origFileName);
-            DEBUG_ASSERT(m_tempFileName.isNull());
-        }
-    }
-    ~SafelyWritableFile() {
-        cancel();
-    }
-
-    const QString& fileName() const {
-        if (m_tempFileName.isNull()) {
-            // If m_tempFileName has not been initialized then no temporary
-            // copy was requested in the constructor.
-            return m_origFileName;
-        } else {
-            return m_tempFileName;
-        }
-    }
-
-    bool isReady() const {
-        return !fileName().isEmpty();
-    }
-
-    bool commit() {
-        if (m_tempFileName.isNull()) {
-            return true; // nothing to do
-        }
-        QString backupFileName = m_origFileName + kSafelyWritableOrigFileSuffix;
-#ifdef __WINDOWS__
-        // After Mixxx has closed the track file, the indexer or virus scanner
-        // might kick in and fail ReplaceFileW() with a sharing violation when
-        // replacing the original file with the one with the updated metadata.
-        int i = 0;
-        for (; i < kWindowsSharingViolationMaxRetries; ++i) {
-            if (ReplaceFileW(
-                        reinterpret_cast<LPCWSTR>(m_origFileName.utf16()),
-                        reinterpret_cast<LPCWSTR>(m_tempFileName.utf16()),
-                        reinterpret_cast<LPCWSTR>(backupFileName.utf16()),
-                        REPLACEFILE_IGNORE_MERGE_ERRORS | REPLACEFILE_IGNORE_ACL_ERRORS,
-                        nullptr,
-                        nullptr)) {
-                // Success, break retry loop
-                break;
-            } else {
-                DWORD error = GetLastError();
-                switch (error) {
-                case ERROR_UNABLE_TO_MOVE_REPLACEMENT:
-                    // The m_tempFileName file could not be renamed. m_origFileName
-                    // file and m_tempFileName file retain their original file names.
-                    kLogger.critical()
-                            << "Unable to rename replacement file"
-                            << m_tempFileName
-                            << "->"
-                            << m_origFileName;
-                    return false;
-                case ERROR_UNABLE_TO_MOVE_REPLACEMENT_2:
-                    // The m_tempFileName file could not be moved. The m_tempFileName file still exists
-                    // under its original name; however, it has inherited the file streams and
-                    // attributes from the file it is replacing. The m_origFileName file still exists.
-                    kLogger.critical()
-                            << "Unable to move replacement file"
-                            << m_tempFileName
-                            << "->"
-                            << m_origFileName;
-                    return false;
-                case ERROR_UNABLE_TO_REMOVE_REPLACED:
-                    // The replaced file could not be deleted. The replaced and replacement files
-                    // retain their original file names.
-                    kLogger.critical()
-                            << "Unable to remove"
-                            << m_origFileName
-                            << "before replacing by"
-                            << m_tempFileName;
-                    return false;
-                case ERROR_SHARING_VIOLATION:
-                    // The process cannot access the file because it is being used by another process.
-                    kLogger.warning()
-                            << "Unable to replace"
-                            << m_origFileName
-                            << "by"
-                            << m_tempFileName
-                            << "because it is used by another process";
-                    QThread::msleep(kWindowsSharingViolationSleepBeforeNextRetryMillis);
-                    continue; // Retry
-                case ERROR_ACCESS_DENIED:
-                    kLogger.critical()
-                            << "Unable to replace"
-                            << m_origFileName
-                            << "by"
-                            << m_tempFileName
-                            << "Access is denied";
-                    return false;
-                default:
-                    // If any other error is returned, such as ERROR_INVALID_PARAMETER, the replaced
-                    // and replacement files will retain their original file names. In this scenario,
-                    // a backup file does not exist and it is not guaranteed that the replacement file
-                    // will have inherited all of the attributes and streams of the replaced file.
-                    kLogger.critical()
-                            << "Error"
-                            << error
-                            << "during replacing"
-                            << m_origFileName
-                            << "by"
-                            << m_tempFileName;
-                    return false;
-                }
-            }
-        }
-        QFile backupFile(backupFileName);
-        if (backupFile.exists()) {
-            if (!backupFile.remove()) {
-                kLogger.warning()
-                        << backupFile.errorString()
-                        << "- Failed to remove backup file after writing:"
-                        << backupFile.fileName();
-                return false;
-            }
-        }
-        if (i >= kWindowsSharingViolationMaxRetries) {
-            // We have given up after the maximum retries in the loop above.
-            return false;
-        }
-#else
-        QFile newFile(m_tempFileName);
-        if (!newFile.exists()) {
-            kLogger.warning()
-                    << "Temporary file not found:"
-                    << newFile.fileName();
-            return false;
-        }
-        QFile oldFile(m_origFileName);
-        if (oldFile.exists()) {
-            DEBUG_ASSERT(!QFile::exists(backupFileName)); // very unlikely, otherwise renaming fails
-            if (!oldFile.rename(backupFileName)) {
-                kLogger.critical()
-                        << oldFile.errorString()
-                        << "- Failed to rename the original file for backup before writing:"
-                        << oldFile.fileName()
-                        << "->"
-                        << backupFileName;
-                return false;
-            }
-        }
-        DEBUG_ASSERT(!QFile::exists(m_origFileName));
-        if (!newFile.rename(m_origFileName)) {
-            kLogger.critical()
-                    << newFile.errorString()
-                    << "- Failed to rename temporary file after writing:"
-                    << newFile.fileName()
-                    << "->"
-                    << m_origFileName;
-            if (oldFile.exists()) {
-                // Try to restore the original file
-                if (!oldFile.rename(m_origFileName)) {
-                    // Undo operation failed
-                    kLogger.warning()
-                            << oldFile.errorString()
-                            << "- Both the original and the temporary file are still available:"
-                            << oldFile.fileName()
-                            << newFile.fileName();
-                }
-                return false;
-            }
-        }
-        if (oldFile.exists()) {
-            if (!oldFile.remove()) {
-                kLogger.warning()
-                        << oldFile.errorString()
-                        << "- Failed to remove backup file after writing:"
-                        << oldFile.fileName();
-                return false;
-            }
-        }
-#endif
-        // Prevent any further interaction and file access
-        m_origFileName = QString();
-        m_tempFileName = QString();
-        return true;
-    }
-
-    void cancel() {
-        if (m_tempFileName.isNull()) {
-            return; // nothing to do
-        }
-        QFile tempFile(m_tempFileName);
-        if (tempFile.exists() && !tempFile.remove()) {
-            kLogger.warning()
-                    << tempFile.errorString()
-                    << "- Failed to remove temporary file:"
-                    << m_tempFileName;
-        }
-        // Prevent any further interaction and file access
-        m_origFileName = QString();
-        m_tempFileName = QString();
-    }
-
-  private:
-    QString m_origFileName;
-    QString m_tempFileName;
 };
 
 } // anonymous namespace
